Refactor module imports and reorganize synaptic dynamics classes for clarity
