+++ brainstate/nn/_conv.py
@@ -23,8 +23,8 @@ import jax.numpy as jnp
 
 from brainstate import init, functional
 from brainstate._state import ParamState
-from brainstate.nn._module import Module
 from brainstate.typing import ArrayLike
+from ._module import Module
 
 T = TypeVar('T')
