--- brainstate/nn/_elementwise/__init__.py
@@ -1,22 +1,0 @@
-# Copyright 2024 BDP Ecosystem Limited. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-# ==============================================================================
-
-from ._dropout import *
-from ._dropout import __all__ as dropout_all
-from ._elementwise import *
-from ._elementwise import __all__ as elementwise_all
-
-__all__ = dropout_all + elementwise_all
-del dropout_all, elementwise_all
