--- brainstate/nn/_event/__init__.py
@@ -1,26 +1,0 @@
-# Copyright 2025 BDP Ecosystem Limited. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-# ==============================================================================
-
-# -*- coding: utf-8 -*-
-
-
-from ._fixedprob_mv import EventFixedProb, EventFixedNumConn
-from ._linear_mv import EventLinear
-
-__all__ = [
-    'EventLinear',
-    'EventFixedProb',
-    'EventFixedNumConn',
-]
