--- brainstate/nn/_interaction/__init__.py
@@ -1,41 +1,0 @@
-# Copyright 2024 BDP Ecosystem Limited. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-# ==============================================================================
-
-from ._conv import *
-from ._conv import __all__ as conv_all
-from ._embedding import *
-from ._embedding import __all__ as embed_all
-from ._linear import *
-from ._linear import __all__ as linear_all
-from ._normalizations import *
-from ._normalizations import __all__ as normalizations_all
-from ._poolings import *
-from ._poolings import __all__ as poolings_all
-
-__all__ = (
-    conv_all +
-    linear_all +
-    normalizations_all +
-    poolings_all +
-    embed_all
-)
-
-del (
-    conv_all,
-    linear_all,
-    normalizations_all,
-    poolings_all,
-    embed_all
-)
