+++ brainstate/nn/_linear_mv.py
@@ -21,8 +21,8 @@ import jax
 from brainstate import init
 from brainstate._compatible_import import brainevent
 from brainstate._state import ParamState
-from brainstate.nn._module import Module
 from brainstate.typing import Size, ArrayLike
+from ._module import Module
 
 __all__ = [
     'EventLinear',
